Create findTwoSumToTarget.dart
You are given a List of integers and a target integer. Write a Dart function that finds and returns the two integers in the list that add up to the target integer. If there are no such integers in the list, the function should return an empty list.

INPUT:
List: [3, 1, 5, 7, 9, 2]
Target: 10
Output:
[1, 9]